backend_service/requirements.txt
kafka-python
psycopg2-binary
streamlit